address_space: fix the bug that other user-space process cannot use shared memory of stratovirt process
The reason is that created File's lifecycle ends when exiting `create_host_mmaps` function,
and the file descriptor which is saved in `HostMemMapping` is invalid after
exiting `create_host_mmaps` function.

This commit fix this bug by saving `Arc<File>` in `HostMemMapping`,
so we can get the file descriptor of this file in the whole lifecycle of
`HostMemMapping` (i.e. the whole lifecycle of stratovirt process).

Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>