clean environment: complete exiting mechanism.
1. Using unrealize function makes code looks elegant, but what if panic
or siginal received occurs when the thread is holding a device lock? deadlock,
right? Because to call unrealize function the thread must hold the block of
the device. To solve this issue, the work of removing files is done by
saving the path of these files in the global variable EventLoop, and removing
them from the variable.

2. Replace register_signal function from libc with register_signal
function by vmm_sys_util.

3. Add handler for SIGSYS signal.

Signed-off-by: Ming Yang <yangming73@huawei.com>