fwcfg: Implement FwCfgFile entry
The FwCfgFile structure is used to enumerate firmware file item from
the directory. It is selected by a 32-bit big-endian number which is
the number of files in the directory. Iterate the directory one by
one until you find a structure with the desired file name. Then populate
the file data using DMA.

Signed-off-by: Ying Fang <fangying1@huawei.com>