virtio: add reset implementation for block device
After probing and using block device, UEFI will set device status to zero.
If we do not reset device at this point, kernel will faile to probe this
device again because it checks the device's status as already activated.

This commit fixes this bug by reseting the device when receiving
device-status-writing command (the command is to set device status to zero)
from UEFI.

Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>