util: extract set_raw/canon_mode to util crate
When launching VM, we will set stdin to raw-mode. And when VM is destroyed,
we will recover stdin to canonical mode. If using `set_raw/conon_mode` function
of vmm-sys-util third-party crate, we can not input `Enter` in UEFI shell.

This commit fixes this problem by implementing seting/recovering terminal mode
in util crate. This commit also defers setting to raw-mode to realization of
PL011/serial devices.

In this commit we save the old mode of terminal in global variable, because we
change the terminal to raw mode when launching a VM and have to recover terminal
to old mode which is saved in global varible. This can gurantee that the mode of
terminal remains unchanged.

Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>