virtio: bugfix of aio process of virtio block device
When configuring Block device's as `-drive file=xxx,direct=on`, Block
device will use aio when processing request. But current aio process can
not handle the situation when the data in request is not aligned.

However, when UEFI send a request via virtqueue, the data is possibly not
aligned and then Block device will return error.

This commit fixes this problem and uses bounce buffer to process not0-aligned
data.

Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>