util: fix log format error when printing log info to stdio
The bug is caused by commit 5930116d. When launching VM configuring `-D`, the
log is redirected to terminal. But the format of StratoVirt's log is abnormal,
and the next line always start at the end-offset of the previous line.
The log looks like follows,

```
INFO: Initrd is configured
                          INFO: vcpu0 is running
```

This problem is relavant to "OPOST" flag of terminal. This flag deactivates
default output processing, in which terminal will automatically insert a
<CR> after receiving a <LF>.

This commit adds this flag to disable the default processing of terminal and
let terminal just print received data.

Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>