util/fdt: implement basic fdt API
This series of patches is for rust update to version >= 1.47.
Because the higher version of rust compile Stratovirt with
aarch64-unkonwn-linux-musl target, will get error like below:

```
  note: /usr/lib/gcc/aarch64-linux-gnu/libfdt.a(fdt_sw.o): In function `fdt_property':
          (.text+0x4d4): undefined reference to `__stack_chk_guard'
          (.text+0x4e0): undefined reference to `__stack_chk_guard'
          (.text+0x518): undefined reference to `__stack_chk_guard'
          (.text+0x530): undefined reference to `__stack_chk_fail'
```

Rust community recommand to use vm-fdt(https://github.com/rust-vmm/vm-fdt)
to fix this problem. But this project is not released to crate.io for now.
So add basic implementation of fdt API.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>