legacy/pl011: fix bug when pasting more than 16 characters
When reading data from standard input, if the number of
characters processed at one time exceeds the available
space in the rfifo vector, the first few characters will
be covered by the later characters. It will cause bug when
pasting more the 16 characters.

Fix it by calculating the size of available space in
rfifo vector, then only receive and process the input
according to the size number.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>
Signed-off-by: Xiaohe Yang <yangxiaohe2@huawei.com>