devices: fix the bug when pasting string more then 16 characters
When pasting the string on aarch64 platform, if the length of pasted string is larger then
32 characters, the program will panic and exit immediately. If the length of
the pasted string is from 16 to 32 characters, a few characters of string will
be overwritten by the last part of the string. This bug is caused by lacking
enough remain space of the receiver buffer when handling the input data in pl011.

The bug is fixed by adding a callback function can_receive_cb() into the
InputReceiver trait which helps handle the input data and ensure the length of
data read at once will not exceed the avaible space.

Signed-off-by: Qi Xi <xiqi1@huawei.com>