syscall: add syscall "newfstatat" in x86_64-unknown-linux-gnu target
When we run stratovirt, built with GNU toolchain on x86 platform, and
use serial port for IO, it will be blocked. The bug is caused by when
it is the first time for rtc device to call "libc::gmtime_r" after
seccomp taking effect, it will use the syscall "newfstatat" which is
not in the syscall whitelist.

The bug is fixed by adding "newfstatat" to syscall whitelist with
x86_64-unknown-linux-gnu target.

Signed-off-by: Qi Xi <xiqi1@huawei.com>