region: fix bug of region compare
When comparing two memory regions, you cannot only compare the
attribute values of the memory region. Because there may be two
overlapping memory areas with identical attribute values.
The correct approach should be to compare the pointers of these
two memory region to ensure that the correct memory region will
be added or deleted.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>