machine/standard_vm: add psci descrition in FADT
PSCI is Power State Coordination Interface on arm plantform, which
is for KVM to handle vm shutdown and reboot. This needs compliant
with hardware reduce bit in FADT flags. But on x86 plantfrom, the
shutdown and reboot of vm rely on ACPI which don't need hardware
reduce bit. Therefore, architecture macros are used to configure
flags respectively.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>