virtio: enhance the performance of net when sending data.
We used to use read function to read date from various vectors and
put them in a big buffer, followed by write them to tap. This is a
usual method, but there is a better method for this scenario: writev.
This function can help to write incontinuous data from serival vectors
into contious buffer without put them in a big buffer in advance.

In this commit, the new function writev is used to improve the performance.

Signed-off-by: Ming Yang <yangming73@huawei.com>