util/loop_context: add two operations for epoll.
Currently there are only two operations offered for users:
AddShared and Delete. But what if a event only should be parked
instead of being deleted and a event only should be resumed from
it was parked instead of being added again? In this commit, the
Park and Resume operations are provided to solve this problem.

When a event is parked, the mapping within eventfd and handler
will still be saved, but the eventfd will not be listened. While
the event is resumed the eventfd is add to epoll again, and it
will be listened again.

Signed-off-by: Ming Yang <yangming73@huawei.com>