virtio/net: use a new strategy for receiving packages.
StratoVirt used to receive packages when both tap_fd and rx_fd are
triggered. However, when the rx virtio queue is full, in other words
there is no buffer to use, StratoVirt still tried to read tap. In this
commit the strategy was changed: When the rx virtio queue is full, tap_fd
will be delete from epoll, when there are buffers available, add tap_fd
to epoll again. In this way, stratovirt receives packages faster.

Signed-off-by: Ming Yang <yangming73@huawei.com>