virtio/queue: make cache for virtio ring to enhance the performance of virtio devices
Whenever an elem poped from the virtio ring, it would be checked.
In the processing of checking, the stratovirt checks if the memory
segment is in some region. So it finds region by region. While in most
cases the memory segment is in the same region, therefore there is no
need to find it every time. So, making a cache for it makes the checking
process works faster.

Signed-off-by: Ming Yang <yangming73@huawei.com>