virtio/block: add io_poll flag to event notifier in virtio-blk
The io_poll flag allows the poll functions to be pre-polled before
each call to epoll_pwait, which greatly improves performance and
avoids the CPU consumption of full polling. In addition, a mut
reference parameter needs to be added in poll functions to determine
whether the polling target function is polling successful or not.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>