virtio/net: fix delete unregister fd error
When a standard VM starts from a raw disk, Stratovir needs to reset
the devices. During this process, the fd of the previously registered
tap device needs to be deleted. However, in the activate function, the
previously registered tap_fd is belong to the original tap device, and
the deleted tap_fd is belong to the tap device which is copied from the
original tap device. An error is reported when the deletion is due to
different fd. Fix it by setting the correct fd.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>