virtio/queue: save host address as cache to enhance the performance of virtio devices
1. Save host address of desc_table, avail_ring and used_ring as cache, When read or
write an object to them, using the host address eliminates the time used by transforming
GPA to HVA.
2. As cache makes virtio device works faster, it is necessary to sync data by fence
function:
If the feature VIRTIO_F_RING_EVENT_IDX is used, devices send notifies by compare the event_idx
and last_used_idx. Normally, the event_idx should equal to the last_used_idx. But, the event_idx
might be the value that the kernel write last time, which means the event_idx is less than the
last_used_idx, which leads to the result, devices do not send interrupt to kernel, devices do
not work any more. In order to address this problem, it it necessary to sync data to let event_idx
is new.
3. Modify test cases.

Signed-off-by: Ming Yang <yangming73@huawei.com>