address_space: add memory prealloc for stratovirt.
Prealloc physical address for virtual machine to improve
the performance of memory when using.

To prealloc memory, StratoVirt read data from mmaped address
and then write them back. To improve the rate of preallocing,
StratoVirt uses multiple threads to touch pages. The number of
threads is the minimum of 16, the number of vCPU and the number
of host CPU.

Signed-off-by: Ming Yang <yangming73@huawei.com>
Signed-off-by: LiChang Zhao <zhaolichang@huawei.com>