vfio: implement Listener trait for struct VfioContainer
Instead of putting all memory regions which are going to be mapped for
DMA into struct VfioMemInfo, we can implement Listener trait for struct
VfioContainer directly. Then DMA mapping and unmapping can be done when
region is added and deleted. In this commit, struct VfioMemInfo is
removed, and method vfio_dma_unmap is added.

Signed-off-by: Jie Yang <yangjieyj.yang@huawei.com>