virtio/console: fix wrong waiting for socket
The console device has a listener fd to monitor whether a socket
initiates a connection. After connection with the socket, the console
device needs to park the listener fd and enable the stream fd to
receive data sent by the socket.

The original reset function does not consider this situation.
As a result, the system waits for the socket connection which cause
main loop stuck. Fixing it by adding an option stream_fd attribute
for the chardev device to determine whether to remove the stream fd.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>