signal_handler: fix deadlock if `handle_signal_sys()` is executed
It may cause deadlock If we print error message in handle signal
function.
How to reoccur: run a vm with `-D`, don't use `-disable-seccomp` cmd.
When running a system call does not appear in the whitelist, a deadlock
occurs.
Why happened: from the stack information, it calls `malloc()` during log
printing. But `handle_signal_sys()` is called when StratoVirt refuses to
execute system calls. In this function, it needs to print error log,
`malloc()` is called again and deadlock occurs.

Signed-off-by: Xinle.Guo <guoxinle1@huawei.com>