migration: use device id as snapshot id
At present, snapshots use the device initialization sequence as the key,
but after the introduction of the device hot-plug mechanism, due to the
dynamic addition of devices, the sequence of snapshot restoration may be
in consistent with the original sequence, resulting in abnormal device
state restoration. Therefore, a new interface is provided, which can
specify the device id during registration, so as to ensure that the
snapshot is restored to match the original device.
In addition, an unregister interface has been added for device
destruction to clean up related resources.

Signed-off-by: zhouli57 <zhouli57@huawei.com>