pci/src: Add parameter for reset in PciDevOps trait
The PciDevOps trait invokes the `restart` function of the corresponding
virtio device through `reset` to reset the device. And for root port, it
also need to reset the device attached in sec_bus during reboot. So add
a extra parameter to determine reset device itself or reset all child
devices.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>