vfio: fix the issue that fails to probe vfio network device
If we start VM with a root port device, and attach the vfio network
device to the root port, the device may fail to activate.
The reason is that in some case, OVMF tries to size the VF BARs and
comes up with absurd results. So we simply hide the SR-IOV extended
capability of pci space for now.

Signed-off-by: Xinle.Guo <guoxinle1@huawei.com>