machine_manager: use write! in TempCleaner::clean
When using info! and other macros to print the log, a lock will be held.
During the processing, if the called function (such as display_chain)
triggers the system signal, the handle_signal_sys function will be
called, and the info! macro will also be called in the
TempCleaner::clean function to print the log, may cause deadlock.

Signed-off-by: zhouli57 <zhouli57@huawei.com>