Fix incorrect use of from_raw_parts_mut function in trait ByteCode.
pub unsafe fn from_raw_parts<'a, T>(data: *const T, len: usize) -> &'a [T]
The function of from_raw_parts is to Forms a slice from a pointer and a length.
The len argument is the number of elements, not the number of bytes.
data must be valid for reads for len * mem::size_of::<T>() many bytes, and it must be properly aligned.

The original code incorrectly used from_raw_parts in the from_bytes function.
The correct use is that the second parameter of len should equal to 1 in the from_raw_parts function.
like this.: let obj_array = unsafe { from_raw_parts::<Self>(data.as_ptr() as *const _, 1) };

Signed-off-by: Yan Wen <yan_ace62@126.com>