standard_vm/x86_64: add default bootorder information to FwCfg device
The Fwcfg device saves information about the boot order, and when the
blk device with the boot order information is added, it needs to be
updated. There is no default boot order information on x86 platforms,
so an error occurs when updating. Add the corresponding default
boot order information to fix the problem.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>