queue_number: Increase the max number of queue config state.
Increase the max number of queue config state to 32. For snapshot
multiqueue virtio device, the `queues_config` number in
`VirtioPciState` struct should be equal to the max value of multiple
queue number. If not ,it will case paninc because the array is out
of range.
refer: https://gitee.com/openeuler/stratovirt/pulls/479

Signed-off-by: Xinle.Guo <guoxinle1@huawei.com>