virtio-gpu: create unready queue for virtio-device
For virtio devices with multiple queues it is a good idea to
alloc multiple queues for the device, even if the driver does
not enable these queues. Because some devices like virtio-gpu
distinguish specific queues, and we use index in vector to
distinguish them.
After this patch, check whether the queue is valid before using it.