VNC: Modify Cargo.toml of vnc and add serveral methods for updating the vnc display
Ramfb is one of the devices that depend on vnc. Delete
devices dependency in Cargo.toml of vnc to avoid circular dependency.

For some display devices like ramfb, it depends on the timer to update
the vnc display. Add a method of using a timer to cyclically update
the vnc display for display devices and a method of getting intervals
to decide when to update the next frame image.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>