virtio-scsi: Add emulation scsi command
We need to emulate some scsi command in stratovirt when using VRDM virtio-scsi.
1. RESUEST SENSE.
The REQUEST SENSE command requests that the device server transfer patameter data
containing sense data to the application client.
2. TEST UNIT READY
The TEST UNIT READY command provides a means to check if the logical unit is ready.
3. INQUIRY
The INQUIRY command requests that information regarding the logical unit and scsi
target device be sent to the application client.
4. READ_CAPACITY_10 (SCSI BLOCK COMMANDS)
The READ CAPACITY(10) command requests that the device server transfer eight bytes
of parameter data describing the capacity and medium format of the direct access
block device to the Data-In Buffer.
5. MODE_SENSE / MODE_SENSE_10
The MODE SENSE command provides a means for a device server to report parameters
to an application client.
6. REPORT_LUNS
The REPORT LUNS command requests that the peripheral device logical unit inventory
accessible to the I_T nexus be sent to the application client.
7. SERVICE_ACTION_IN_16
8. WRITE_SAME_10 / WRITE_SAME_16
9. SYNCHRONIZE_CACHE

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>