Fix: when the vcpu stopped, don't kick it
After the VM is shut down, one of the vcpus exits the user mode and
this vcpu status changes to stopped. First, in the vcpu.destroy function,
this vcpu wastes 32 ms waiting time. Second, this vcpu sends a signal
to itself and may end the vcpu thread advance. As a result, other vcpu
do not perform the kick operation and the VM is suspended.

Signed-off-by: Mingwang Li <limingwang@huawei.com>