Fix: don't realize vsock device when VM reboot
The fd is configured for the vsock device. Therefore, the fd of the
devices is not released when the VM is restarted. If the reset operation
is performed on the device, the fd.set_owner command is executed repeatedly.
The kernle don't allow the set_owner command to be executed repeatedly.
As a result, the VM fails to restarted.

Therefor, when the VM is restarted, it only need to suspend the device.

Fix me: The device queue may need to perform corresponding operations.

Signed-off-by: Mingwang Li <limingwang@huawei.com>