virtio-console: Fix some problems of using virtio-console
The console which users use the socket file or the "virsh console" command
to log in will be stuck after reboot. It is due to the fact that the
listening event of UnixStream connection between guest and virtio-console
was deleted improperly. It also causes that the VM will be killed after
reboot twice and the VM will fail to boot if users enter any character in
console immediately after reboot.

Add a console_connected status to indicate to the connectability between
guest and console.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>