virtio-scsi: Realize deactivate function
VirtioDevice function will be called during vm starts as such order:
First activate() -> deactivate() -> Second activate().

Now, deactivate function is not implement. As a result of that, EventFd
requested in the first activate() will not be freed, and read it will report
error after requesting a new EventFd in the second activate(). So,
realize deactivate() function to delete the old EventFd.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>