VmConfig/object: Modify `object` element in VmConfig
For all about VM's features command line `-object xxx`, the
parameters are saved in VmConfig struct object element.
If the features ID are the same, the parameter will be overwritten.

For example:
1. For NUMA node: -object memory-backend-ram,id=tmp,size=xxx;
2. For RND device: -object rng-random,id=tmp,xxx;
One of the parameters will be overwritten as above.

How to slove: add new struct `ObjectConfig`, and add an element
for each type of parameter.
```
pub struct ObjectConfig {
    pub rng_object: ...,
    pub mem_object: ...,
    ...
}
```

Signed-off-by: Xinle.Guo <guoxinle1@huawei.com>