pci: notify only when the enabled/masked bits are written
Only when the masked/enabled bits in msix cap configure space is written,
it's necessary to check whether to notify the guest. Otherwise, let the
msix/pba table written callback functions to deal with that, or just leave
it alone, as that such as writing cap next register should not affect the
interruption behavior.

Signed-off-by: Zhang Bo <oscar.zhangbo@huawei.com>