vhost-net: using set_backend() to reset device
In guest, execute "reboot" command, it will call reset() in
StratoVirt. The taps will be assigned None value, which will
close the tap's fd later. When it call realize() later, it
will use the fd which has been closed. So, it cause the reboot
process gets stuck.

Instead of setting taps to None, we use set_backend() with fd(-1)
to reset device.

Signed-off-by: Yan Wang <wangyan122@huawei.com>