virtio-mmio: recreate interrupt_status cause snapshot getting stuck
Fix commit 5d8eb48 "virtio: add report_virtio_error()".

After commit 5d8eb48, the interrupt_cb in VirtioMmioDevice::activate and
VirtioMmioDevice::resume is unified to the same one in VirtioMmioDevice::realize.

This will cause the vm getting stucked when restoring it from the snapshot.
The stucking process:
realize()
  ->assign_interrupt_cb
    ->let interrupt_status = self.interrupt_status.clone();
    ->cb = Arc::new(Box::new(..)); // In cb, it will use interrupt_status.
    ->self.interrupt_cb = Some(cb);
set_state_mut()
  ->self.state = Arc::new(..);
  ->self.interrupt_status = Arc::new(...); // This will cause the old clone meaningless.
resume()
  ->activate(.., self.interrupt_cb,..)
  ...
  ->self.interrupt_cb()
    ->interrput_status.fetch_or(..); // It is old clone, "fetch_or" is meaningless.
    ->interrupt_evt.write(1)...; // The interrupt doing nothing due to the old cloned interrupt_status.

So, the interrupt for vm is not processed correctly, the vm will be stucked.
We should correct the set_state_mut(), not release the reference on old value for "state"
and "interrupt_stats, just modify its value.

Signed-off-by: Yan Wang <wangyan122@huawei.com>