virtiofsd: Report error when the client queue request was in polling mode
For VHOST_USER_SET_VRING_KICK and VHOST_USER_SET_VRING_CALL and VHOST_USER_SET_VRING_ERR,
Bits (0-7) of the payload contain the vring index. Bit 8 is the invalid FD flag. This flag
is set when there is no file descriptor in the ancillary data. This signals that polling
should be used instead of waiting for the kick.
We check this flag, and report error when the client queue request was in polling mode.

Signed-off-by: Li HuaChao <lihuachao1@huawei.com>