pci: do not report error in some scenarios
Because linux kernel tries to read offset 256B of the configuration
space of the host bridge, although it's a PCI device that has only
256B meaningful space(that's to say, the max offset is 255B). Meanwhile,
virtio-net driver doesn't set the interrupt vector for the control queue,
thus, we would use the default 0xFFFF vector to notify the guest, which
is invalid, and an error message occurs. These abnormal behaviors are
triggered by the guest, thus tolerate them by changing loglevel from error
to warn.

Signed-off-by: Zhang Bo <oscar.zhangbo@huawei.com>