virtio-blk: Add more limitation to IO merging
Linux limits the size of iovecs to 1024 (UIO_MAXIOV in the kernel
sources, IOV_MAX in POSIX). Because of this, merged blk requests
with many iovecs are rejected with -EINVAL by the io_submit() or
readv()/writev() system calls.

And refer to Qemu that has a limitation called max_transfer size
of a single request to i32::MAX, which has some complex context
reason. To be conservative, let's add this limitation too.

Signed-off-by: Keqian Zhu <zhukeqian1@huawei.com>