virtiofs: introduce FUSE_IOCTL to virtiofsd
Normally the VM should not use ioctl to modify files, but it can be useful in some
cases. For example: to modify inode attrs, it is required for per inode DAX. We do
not currently support these features, so ioctl is not currently required. Reserve
this interface and implement it when needed in the future.

Signed-off-by: Li HuaChao <lihuachao1@huawei.com>