Fix: Each queue is assigned a deactivate eventfd.
VM startup consists of two processes: EDK startup and guest kernel
startup. During EDK startup, the driver enables the multi-queue of the
Blk device and then destroys the multi-queue. Destroying multi-queue is
an asynchronous operation and a broadcase operation. As a result, the
guest kernel initializes the Blk device, and the old resoureces are not
released. Some queues use the information in the EDK phase, triggering
interrupt loss.

Now configure a unique eventfd for each queue deactivated operation. The
ioeventfd deregistration mechanism will be added later to prevent the
guest kernel from using old resources.

Signed-off-by: Mingwang Li <limingwang@huawei.com>