usb: only the do_parameter needs to be called on the control plane
According to the spec 4.11.2.2, a Setup Stage TRB shall contain
immediate data(IDT flag = '1'), its Parameter fields shall contain the
8-byte USB SETUP Data, which defines the request and the request's
parameters that will be sent to the device in the USB Setup stage
transaction, and its Length field shall be set to '8'.
Therefore, only the do_parameter need to be called.

Signed-off-by: zhouli57 <zhouli57@huawei.com>