xhci: simplify the submit transfer
According to the spec, the following conditions should force Transfer
Event generation to take place immediately:

The completion of a TRB that has its IOC flag set.

The completion of a Short Packet on a TRB that has its ISP flag set.

An error occurs on any Treansfer TRB.

We use submit_transfer to report succeed transfer, and use
report_transfer_error to report error.

Signed-off-by: zhouli57 <zhouli57@huawei.com>