address_sapce: check whether a memregion has fd changed when update ioeventfds
In the last patch we make eventfd unchanged by stratovirt in the memory part,
thus we can now tell whether the fd related memory region is updated by the
guest or not. If the guest has update_bar_mappings and makes the map between
mem address and eventfd different with BIOS(it's true for linux guestes), we
can now tell. If that happens, we re-set the listener for that fd.

Signed-off-by: Zhang Bo <oscar.zhangbo@huawei.com>