Pci: Fix the secondary bus reset problem
According to the PCIe specification, setting the BRIDGE_CTL_SEC_BUS_RESET bit
of the root port configuration triggers a hot reset the corresponding root port.
The operation was missing before. So fix it.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>

Pci: Fix the secondary bus reset problem

According to the PCIe specification, setting the BRIDGE_CTL_SEC_BUS_RESET bit
of the root port configuration triggers a hot reset the corresponding root port.
The operation was missing before. So fix it.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>