demo_dev: add the data handling bar
If the mmio address is writen, multiply the data[0] with 2. The data is
reserved in a hashmap that's inside this device. The data will be removed
after reading, as that we think it becomes useless after the test process
has validated the data.
Meanwhile, rm init_bars() because bars should be registered at the time
that we register region_ops.