memory: change high pcie mmio range for PCI host
The high pcie mmio base should get the same alignment as the size,
ie. a 512GiB entry will be aligned on a 512GiB boundary. So we just
change the base to 512 GiB. At the same time we also modify
accordingly HighGicRedist and HighPcieEcam.
We also need to make modifications to avoid overlapping the MEM area
with the PCIe area at the back.

Signed-off-by: Li HuaChao <lihuachao1@huawei.com>