MST: bugfix scream playback basic test occasional failed
When the value of fmt is assigned after is_started, the back-end directly
obtains the value of fmt after judging is_started. At this time, the
front-end fmt is still in the process of assignment, and the obtained
value is not the final value. When the backend checks fmt, the fmt is
changed, and the chunk_idx value is reassigned. The data in the first
frame is lost.

So, the is_started must be setting at the end.

Signed-off-by: Mingwang Li <limingwang@huawei.com>