PCI-MST: Add two MST testcases for INTx
Add two MST testcases. One testcase is for testing the basic function of INTx.
The other testcase is for testing the interrupt disable bit of the command
register that is used to control the ability of a Function to generate INTx
emulation interrups.

Add some functions to simulate interrupt triggering and EOI.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>