camera: performance: enlarge MAX_PAYLOAD to reduce datacopy times
According to USB video FAQ, "MaxPayloadTransferSize should be set to a value comprised
between 10ms worth of data and dwMaxVideoFrameSize.", thus we set the MAX_PAYLOAD to
typical uncompressed video frame size, which is 1024 * 768 * 2. Note that MJPG do not
have fixed frame size, thus this number shall be according to expierience. 1024 * 768 * 2
seems a sensible trade-off.

Signed-off-by: Zhang Bo <oscar.zhangbo@huawei.com>