usb-host: Add realize, unrealize and related methods.
Locate the device from host based on one of the parameter combination of
hostbus and hostaddr, hostport and vendorid and productid. And detach the
USB device from host in the realizing process.

Release the interfaces of the host USB device and attach the USB device
to host in the unrealizing process.

Signed-off-by: Mingwang Li <limingwang@huawei.com>