migrate: use serde instead of unsafe call during snapshot restore
As byte_order in MigrationHeader is set to Niche if it's 0 in snapshot file,
due to Discriminant Elision, and could not be deferenced. Thus using
from_bytes() to convert bytes to MigrationHeader will deference it, and cause
an error. Using Ok() to wrap it it will change Ok() into Error (on Rust 1.67
and above), As Ok() and Err() share the same memory for their members, the
ErrorVTable in Error is overriden by MigrationHeader which is wrapped by Ok().
To avoid such unsafe behavior, remove the unsafe as_ref()/from_bytes() to
remove the problem from the root, and use serde to help to avoid further problems.

Signed-off-by: Zhang Bo <oscar.zhangbo@huawei.com>