Scream: Optimize the playback and capture process
Add a destroy method for the AudioInterface trait. Destroy the stream
of playback or capture when it ends. Persistently send or receive chunks
until the stream ends instead of cyclically send or receive a chunk
after checking whether the stream is ready.
The benefit of modification is playing/capture the data as more as
possible. It avoids checking whether the stream is ready after play/
capture a chunk of data every time.

Signed-off-by: Jinhao Gao <gaojinhao@huawei.com>