devices: fixed data drop in legacy serial
There was an issue that method Serial::remain_size returns capacity of a buffer instead of free space in it while method Serial::receive checks that the buffer has enough space to place data into it. Due to this inconsistency client could pass into Serial::receive method a data which size exceeds available space in serial's buffer and as a result this data be dropped.

Signed-off-by: prostous <prostov.yury@huawei.com>
Signed-off-by: Keqian Zhu <zhukeqian1@huawei.com>