usb: change the USB control request log level to warn
Guest can maliciously inject incorrect requests to cause the
backend to print logs crazily. Due to this being a serious
error, we must keep this log for the purpose of locating the
issue. Change the log level to warn, and since logs can be
sharded and dumped, the risk of backend log explosion is
mitigated by other components.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>