Allow efi-bios size be less then memory-region
we check that if pflash-param is read-only. And if so,
we allow size smaller then specified memory region size.

We fail if file-size is zero or if size is greater then
specified region size.

This allows to boot using EFI-Bios image with size less then 64M
without padding it with zeroes

Signed-off-by: Dmitry Skorodumov <skorodumov.dmitry@huawei.com>