timer: Fix possible accident removal of timer
Use memory location to locate a timer is unsafe, as the
old timer may has been expired and a new timer use the same
memory location, so when we delete a timer, the new timer
is deleted.

This commit use incremental number to identify a timer and
u64 is big enough for use.

Signed-off-by: Keqian Zhu <zhukeqian1@huawei.com>