Coding style: standardize the use of `Result`
Most `Result` in the code are from the `Anyhow` lib which is widely used in code.
There is no need to introduce the `Result` in mod.rs and then use it in xxx.rs
by using `$mod_name::Result` and renaming it.
Meanwhile, it does not require `pub use`.

Fix these problems and standardize the use of `Result`.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>