virtio-net: call handle_rx after queue is not full
Due to the EDGE_TIGGERED for tap device, if queue is full and left some
packets in the tap device, it may not trigger handle_rx again. These packets
may not be processed.
So, it shoule call handle_rx to read packet from the tap device after the queue
is not full.

Signed-off-by: Yan Wang <wangyan122@huawei.com>