Implement RX Flow Control for serial ports
Before the patch, when Serial-port is not ready
to accept data, we paused all RX activity on port
for 50ms. It's turned out, that this noticeably slows
down serial port. For example, on pl011 FIFO is 16 bytes.
with this approach pl011 is able to accept only 320 bytes/second
and WinDBG doesn't work

Implemented some flow-control for RX. When chardev receives
external data and serial::remain_size() returns zero,
chardev notifies serial that RX is paused. And removes
IN-event from ppoll() of chardev-file-descriptor.

When Serial sees that new space appeared and it can
accept new data again, serial calls chardev::unpause_rx()
and IN-event is again added to ppoll().

For uart and pl011 we just see that data was read from fifo.

For virtio-serial, we subscribe for input-queue notifications
and see when new input-buf appears.

Test shows (contact me for virtio-serial performance measurement
utility), that Host->Guest transmission performance is

Release build:
- After the patch: about 250 (sometimes 320) MByte/Second
- before patch: It works very slow. about 25 (twenty five) MByte/Second

If test-program is slowed down by generating random-numbers
for each buffer, then situation is a bit better:
- After the patch: about 135 (up to 160) MByte/Second
- before patch: about 120 (130 is maximum I saw) MByte/Second

Signed-off-by: Dmitry Skorodumov <skorodumov.dmitry@huawei.com>